+++ src/components/scripts/kubernetes/using_namespace.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+namespace="$(kubectl config view --minify --output 'jsonpath={..namespace}'; echo)"
+
+if [ -n "${namespace}" ]; then
+    echo -e "\nUsing namespace $(kubectl config view --minify --output 'jsonpath={..namespace}'; echo)"
+fi